Address custom image terminal helper duplication
